Features:
Installed Maatwebsite\Excel and Spatie\ArrayToXml to enable the user to export and download the data with
author, title, or both fields as CSV or XML file.

Created the edit.blade.php UI so that the user can edit a book'a author.

BooksController.php: Created an export method to allow user to export data with the help of Excel library.

config/app.php: Added the dependencies for excel library.

routes/web.php: Implemented a route to the export method in BooksController.php.

index.blade.php: Added padding between the search bar and the table.

edit.blade.php: Created the UI for the edit route.